Basic IR changes for first-class inferred ports
